feat: Move device editing to context menu
Device editing and removal functionality has been moved from the `DeviceCard` component to the context menu. This provides a more consistent and less intrusive way to manage devices, especially when not in edit mode. The `isEditMode` check has also been removed from the context menu handler as it is no longer necessary.